+++ cinemaAutoatendimento/src/main/java/com/example/cinemaAutoatendimento/repository/FilmeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmeRepository extends JpaRepository<FilmeModel, Integer> {
+public interface FilmeRepository extends JpaRepository<FilmeModel, Long> {
 }
