+++ cinemaAutoatendimento/src/main/java/com/example/cinemaAutoatendimento/repository/PessoaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PessoaRepository extends JpaRepository<PessoaModel, Integer> {
+public interface PessoaRepository extends JpaRepository<PessoaModel, Long> {
 }
