+++ cinemaAutoatendimento/src/main/java/com/example/cinemaAutoatendimento/service/AssentoService.java
@@ -19,7 +19,7 @@ public class AssentoService {
     @Autowired
     SessaoRepository sessaoRepository;
 
-    public AssentoModel salvarAssento(int id, AssentoModel assento) {
+    public AssentoModel salvarAssento(Long id, AssentoModel assento) {
         SessaoModel sessao = sessaoRepository.findById(id).orElseThrow(() -> new RuntimeException("Sessão não encontrada"));
 
         assento.setSessao(sessao);
@@ -27,29 +27,29 @@ public class AssentoService {
     }
 
 
-    public List<AssentoModel> listarAssentosPorSessao(int sessaoId){
+    public List<AssentoModel> listarAssentosPorSessao(Long sessaoId){
         return assentoRepository.findBySessaoId(sessaoId);
     }
 
-    public List<AssentoModel> listarAssentosLivresPorSessao(int sessaoId){
+    public List<AssentoModel> listarAssentosLivresPorSessao(Long sessaoId){
         return assentoRepository.findBySessaoIdAndPessoaIsNull(sessaoId);
     }
 
-    public AssentoModel ocuparAssento(int id, PessoaModel pessoa){
+    public AssentoModel ocuparAssento(Long id, PessoaModel pessoa){
         AssentoModel assento = assentoRepository.findById(id).orElseThrow(() -> new RuntimeException("Assento não encontrado"));
         assento.setPessoa(pessoa);
 
         return  assentoRepository.save(assento);
     }
 
-    public AssentoModel desocuparAssento(int id) {
+    public AssentoModel desocuparAssento(Long id) {
         AssentoModel assento = assentoRepository.findById(id).orElseThrow(() -> new RuntimeException("Assento não encontrado"));
         assento.setPessoa(null);
 
         return  assentoRepository.save(assento);
     }
 
-    public boolean excluirAssento(int id){
+    public boolean excluirAssento(Long id){
         if (assentoRepository.existsById(id)){
             assentoRepository.deleteById(id);
             return true;
